# openapi.yaml
openapi: 3.0.3
info:
  title: Student AI API
  version: 0.1.0
  description: Contract between the Django backend and the React console / student
    PWA. This schema is the source of truth — regenerate with `manage.py spectacular
    --file ../openapi.yaml` after any serializer change.
paths:
  /api/auth/csrf/:
    get:
      operationId: auth_csrf_retrieve
      description: Sets the `csrftoken` cookie. Call once before POSTing to login.
      tags:
      - auth
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Detail'
          description: ''
...
  /api/me/:
    get:
      operationId: me_retrieve
      description: |-
        Who the caller is, and which institute they are bound to.

        The console reads `role` to decide which navigation to render, and
        `institute` to label it. Both come from the same place the tenant
        middleware reads, so a UI that trusts this payload and a database that
        enforces RLS cannot disagree about who the caller is.
      tags:
      - auth
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Me'
          description: ''
...
components:
  schemas:
...
    Detail:
      type: object
      description: The one-key error/ack envelope DRF already uses everywhere.
      properties:
        detail:
          type: string
      required:
      - detail
...
    Institute:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 200
        city:
          type: string
          maxLength: 100
        slug:
          type: string
          maxLength: 80
          pattern: ^[-a-zA-Z0-9_]+$
      required:
      - city
      - id
      - name
      - slug
...
    Me:
      type: object
      description: The caller's identity. Drives navigation and the institute label.
      properties:
        id:
          type: integer
        username:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          nullable: true
          description: |-
            Null for an authenticated account attached to nothing — which sees no tenant data at all, by design.

            * `mentor` - mentor
            * `student` - student
            * `director` - director
          oneOf:
          - $ref: '#/components/schemas/RoleEnum'
          - $ref: '#/components/schemas/NullEnum'
        institute:
          allOf:
          - $ref: '#/components/schemas/Institute'
          nullable: true
        mentor_id:
          type: integer
          nullable: true
        student_id:
          type: integer
          nullable: true
        is_staff:
          type: boolean
        is_superuser:
          type: boolean
      required:
      - email
      - id
      - institute
      - is_staff
      - is_superuser
      - mentor_id
      - name
      - role
      - student_id
      - username
...
    NullEnum:
      enum:
      - null
...
    RoleEnum:
      enum:
      - mentor
      - student
      - director
      type: string
      description: |-
        * `mentor` - mentor
        * `student` - student
        * `director` - director
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid